# apps/be/public/swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "BE API",
    "version": "1.0.0",
    "description": "Backend API - Auto-generated from JSDoc"
  },
  "servers": [
    {
      "url": "http://localhost:8000",
      "description": "Development"
    }
  ],
  "paths": {
...
    "/api/stacks": {
      "get": {
        "tags": [
          "Stacks"
        ],
        "summary": "Get all stacks with card counts",
        "responses": {
          "200": {
            "description": "List of stacks with card counts",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "type": "object",
                    "properties": {
                      "id": {
                        "type": "integer"
                      },
                      "name": {
                        "type": "string"
                      },
                      "cover": {
                        "type": "object",
                        "properties": {
                          "type": {
                            "type": "string",
                            "enum": [
                              "color",
                              "gradient",
                              "image"
                            ]
                          },
                          "value": {
                            "type": "string"
                          }
                        }
                      },
                      "cardCount": {
                        "type": "integer"
                      },
                      "createdAt": {
                        "type": "string"
                      }
                    }
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "Stacks"
        ],
        "summary": "Create a new stack",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "name"
                ],
                "properties": {
                  "name": {
                    "type": "string"
                  },
                  "cover": {
                    "type": "object",
                    "properties": {
                      "type": {
                        "type": "string",
                        "enum": [
                          "color",
                          "gradient",
                          "image"
                        ]
                      },
                      "value": {
                        "type": "string"
                      }
                    }
                  }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Stack created",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "id": {
                      "type": "integer"
                    },
                    "name": {
                      "type": "string"
                    },
                    "cover": {
                      "type": "object"
                    },
                    "createdAt": {
                      "type": "string"
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
    "/api/stacks/{id}": {
      "get": {
        "tags": [
          "Stacks"
        ],
        "summary": "Get stack by ID with its cards",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Stack with cards",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "id": {
                      "type": "integer"
                    },
                    "name": {
                      "type": "string"
                    },
                    "cover": {
                      "type": "object",
                      "properties": {
                        "type": {
                          "type": "string",
                          "enum": [
                            "color",
                            "gradient",
                            "image"
                          ]
                        },
                        "value": {
                          "type": "string"
                        }
                      }
                    },
                    "cards": {
                      "type": "array",
                      "items": {
                        "type": "object",
                        "properties": {
                          "id": {
                            "type": "integer"
                          },
                          "stackId": {
                            "type": "integer"
                          },
                          "name": {
                            "type": "string"
                          },
                          "cover": {
                            "type": "string"
                          },
                          "description": {
                            "type": "string"
                          },
                          "createdAt": {
                            "type": "string"
                          }
                        }
                      }
                    },
                    "createdAt": {
                      "type": "string"
                    }
                  }
                }
              }
            }
          },
          "404": {
            "description": "Stack not found"
          }
        }
      },
      "put": {
        "tags": [
          "Stacks"
        ],
        "summary": "Update a stack",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "name": {
                    "type": "string"
                  },
                  "cover": {
                    "type": "object",
                    "properties": {
                      "type": {
                        "type": "string",
                        "enum": [
                          "color",
                          "gradient",
                          "image"
                        ]
                      },
                      "value": {
                        "type": "string"
                      }
                    }
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Stack updated"
          },
          "404": {
            "description": "Stack not found"
          }
        }
      },
      "delete": {
        "tags": [
          "Stacks"
        ],
        "summary": "Delete a stack and all its cards",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Stack deleted"
          },
          "404": {
            "description": "Stack not found"
          }
        }
      }
    },
...
  },
...
}